bin/save-ebuild-env.sh: refrain from using the compgen builtin
For the compgen builtin to be available requires that bash be compiled
with --enable-readline. Rather than rely on compgen to generate a list
of function names, parse their names out of declare -F instead. This
eliminates an unnecessary divergence from the master branch.

Specify -v for the following unset command, so that bash is coerced into
unsetting only variables, as is intended.

Expand the applicable variable names with "${!___@}". Owing to the
constraints placed on identifiers, it's not particularly important that
it be done this way, but I think it better expresses the intent.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Bug: https://bugs.gentoo.org/909148
Signed-off-by: Sam James <sam@gentoo.org>