portage(5): fix description of volatile attribute
The description of the 'volatile' keyword in the 'repos.conf' section in the portage(5)
man page seems to be inconsistent.

I think the first two sentences correctly describe the behaviour of 'volatile' and
agree with the comments in portage/repository/config.py:
"Specifies whether a repository is volatile.  Volatile repositories are assumed to
contain changes made outside of Portage.  This prohibits optimizations from occurring
by assuming the integrity of the repository and its caches may be  violated."

The remainder of the paragraph inverts the meaning of 'volatile' and does not
agree with the code in portage/repository/config.py:
"Disabling this option allows local changes to be made to the repository for
e.g. development safely.  For git-synced repositories, Portage will not perform
destructive operations.  Per default, repositories are considered volatile if
their directory is owned by either 'root' or 'portage' and if it is under
/var/db/repos. Otherwise, the repository is considered non volatile."

Bug: https://bugs.gentoo.org/911594
Signed-off-by: Christian Becke <christianbecke@web.de>
Signed-off-by: Sam James <sam@gentoo.org>