ebuild: inject implicit libc RDEPEND
Inject >=${LIBC_PROVIDER}-${VERSION_OF_LIBC_PROVIDER} into RDEPEND.

We already try to upgrade the virtual/libc provider and its deps aggressively
but that's not so helpful if there's a binpkg for, say, one of its deps available
built against a newer glibc - which ends up breaking the system because our
glibc isn't new enough (symbol versioning).

This is a long-standing problem for binpkg users and one of the last showstoppers
for being able to use them reliably.

Note that this applies to source installs too, although it matters less there;
it'll make downgrading libc a bit harder for people who want to do that, but users
are already prevented from doing that (pkg_* check) for glibc, and I don't really
see it as a bad thing to effectively propagate this to other libcs.

To fully solve the problem, we should arguably do at least one of the following:
1) Implicit >= on anything (not just libc) which installs ELF (at least a PROVIDES);
2) Implement the suggestion in bug #753500 based on analysis of used versioned symbols (*).

But libc is really the critical one and the one where things explode pretty badly,
especially combined with us trying to Do The Right Thing for non-binpkg cases
(aggressively upgrading libc before anything else). The other cases don't matter
so much as they get upgraded later and by that point, the library is usually done.

(It's not really clear if downgrading musl works and even if it is supported,
I'm not really sure it's a supported case at all, so I'm not bothering to carve
out an exception here. It'd make this far less elegant and I don't see any benefit
to doing so.)

(*) util-linux is one of the examples I have in mind here as justification
for either point.

Bug: https://bugs.gentoo.org/753500
Bug: https://bugs.gentoo.org/913628
Signed-off-by: Sam James <sam@gentoo.org>