emerge: warn on installed packages in @world without ebuilds available
This is an outrageously common issue for users: they emerge something
a while ago, it gets treecleaned (possibly without the user ever seeing
the p.mask entry if they don't use the particular system often), but Portage
diligently works to keep the package installed and its dependencies satisfied.

Portage is completely in the right here, but it's very easy for users
to not be aware that this is still a constraint they've asked for that is
being honoured.

This is particularly bad if that package has a slot-operator dependency (:=)
on something which has just changed subslot, as it can't be rebuilt (because
no ebuild), and users get confusing conflicts as a result.

Anyway, for the actual changes:
* Use _replace_installed_atom which is already used for solving slot-operator
conflicts as a proxy for whether we have an ebuild (or something close enough).

* Warn loudly, as we do for other broken things in the world file, if an entry
in there has no ebuilds available.

At one point, I did also have:

* Make _replace_installed_atom's scanning for similar slots optional to speed
things up for our "is an ebuild available for thing in world" check. It remains
on for all existing consumers.

but Zac pointed out this may lead to noise where dependency resolution
would then succeed and indeed I can't convince myself it was right, so
I've dropped that.

Bug: https://bugs.gentoo.org/911180
Signed-off-by: Sam James <sam@gentoo.org>