env-update: Write /usr/etc/ld.so.conf to fix bfd in some obscure cases
This is only needed on prefixed systems. bfd currently reads
${EPREFIX}/etc/ld.so.conf and adds the prefix to these paths, but these
paths are already prefixed. We need them to stay prefixed for the
runtime linker. bfd will use ${EPREFIX}/usr/etc/ld.so.conf instead if
that is present, so we can write the unprefixed paths there.

Other linkers do not use these files at all. We tried to patch bfd to
not use them either, as it shouldn't really be necessary, but that
broke some cases, so we are trying this safer approach instead.

env-update does not write the files under /etc/ld.so.conf.d, but we
shouldn't need to handle these in any case, as all known instances are
not affected by this issue.

Bug: https://bugs.gentoo.org/892549
Signed-off-by: James Le Cuirot <chewi@gentoo.org>