ebuild: do not drop FEATURES=test when USE=-test
By convention the "test" USE flag is used to control dependencies required
for execution of the src_test function.

The "test" USE flag has no special meaning in PMS.

The Gentoo Policy Guide [1] advises that ebuilds must make use of RESTRICT
to disable tests based on USE flags.

Current versions of Portage will not run tests when RESTRICT=test is
set, regardless of the FEATURES setting.

This code in Portage predates the Gentoo policy. It is unneeded since
the policy was created and ebuilds have been updated with the necessary
RESTRICT values.

[1] https://projects.gentoo.org/qa/policy-guide/other-metadata.html#pg0703

Signed-off-by: Mike Gilbert <floppym@gentoo.org>