find_smallest_cycle: Do not return single node for asap mode
For asap mode, it is possible for gather_deps to prematurely
select a single node instead of a cycle, as observed in bug
917259. Continue search for a true cycle in this case.

When not in asap mode, we sometimes need to select a single
node here, as demonstrated by PerlRebuildBugTestCase which
selects automake this way.

The included AlternativesGzipTestCase demonstrates correct
behavior, though it fails to reproduce bug 917259. Even
though this patch has not yet been proven to fix bug 917259,
it is still useful because it limits the ability of
find_smallest_cycle to misbehave.

Bug: https://bugs.gentoo.org/917259
Signed-off-by: Zac Medico <zmedico@gentoo.org>