Fix move_ent with signed binpkg
The gpkg file that cannot be updated will be removed.

[sam: We did discuss stripping the signatures to re-use them but that
feels messier and likely unexpected by the user. We also don't distinguish
between locally-made vs fetched binpkgs (where perhaps for local ones, we could
just re-sign if signed).

It was also raised that for fetched binpkgs, we could not worry once it's
been verified the first time, but that's again complicated.

Simply dropping these binpkgs seems like the best solution and to allow
the binhost to re-sign it on updating instead.]

Bug: https://bugs.gentoo.org/919419
Signed-off-by: Sheng Yu <syu.os@protonmail.com>
Closes: https://github.com/gentoo/portage/pull/1043
Signed-off-by: Sam James <sam@gentoo.org>