Add merge-wait FEATURES setting enabled by default
Add a new merge-wait FEATURES setting, enabled by default,
which prevents packages from being merged while other packages
are building. Previously, this behavior was already implemented
for packages that satisfy direct or indirect dependencies of
the system set. This feature can be disabled only for other
packages which do not satisfy direct or indirect dependencies
of the system set, in order to trade the possibility of random
build failures for greater parallelism.

Currently, it is known that having merge-wait disabled can
cause "random" build failures for builds using setuptools when
setuptools plugins are merged during the build.

Bug: https://bugs.gentoo.org/663324
Signed-off-by: Zac Medico <zmedico@gentoo.org>