save-build-env: Strip out MAKEOPTS and NINJAOPTS
MAKEOPTS and NINJAOPTS contains a host specific value. If the same
package is built on a host with different core counts, the binpkgs will
have different hashes. This change drops MAKEOPTS after src_install so
that it's not included in final environment.gz. This change also adds
them to the special_env_vars.py so that we always have the most up to
date version of these variables.

Bug: https://bugs.gentoo.org/914441
Bug: https://bugs.gentoo.org/736794
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Part-of: https://github.com/gentoo/portage/pull/1311
Signed-off-by: Sam James <sam@gentoo.org>