Do not export A in EAPI 9
Instead of passing A as part of the process environment, we pass it
via a file. Since A is usually the greatest contributor to the process
environment, removing it from the process environment significantly
avoids running into MAX_ARG_STRLEN when spawning a new child process.

This means that A is no longer exported in the ebuild and hence
unavaiable to child processes. However, A is mostly used as part of
the default_src_unpack function and there A does not need to be
exported.

Thanks to Zac Medico for helpful input on this change.

Closes: https://bugs.gentoo.org/721088
Signed-off-by: Florian Schmaus <flow@gentoo.org>