PoC: Do not export A
Instead of passing A as part of the process environment, we pass it
via a file. Since A is usually the greatest contributor to the process
environment, removing it from the process environment significantly
helps running into MAX_ARG_STRLEN when spawning a new child process.

This means that A is now unexported in the ebuild. However, A is
mostly used as part of the default_src_unpack function. And there A
does not need to be exported.

Proof-of-concept for https://bugs.gentoo.org/721088

Bug: https://bugs.gentoo.org/721088
Signed-off-by: Florian Schmaus <flow@gentoo.org>