output.py: try to use curses for setting xterm title
Currently only the list of hard coded terminal names found in
_legal_terms_re allow usage of the title setting feature. Maintaining
this list is a burden and means that some terminals will miss out this
feature, although they actually do have support for it.

Thus, try to use the curses module to dynamically query the terminfo
database for "tsl" (to status line) and "fsl" (from status line)
strings. This way all terminals that declare support for the feature
will automatically benefit.

As a fallback, should curses not be available, still rely on the
_legal_terms_re for setting up terminal titles.

Signed-off-by: Matthias Gerstner <matthias.gerstner@nefkom.net>