bin/ebuild.sh: disable globskipdots too in Bash 5.2
See 69cac73ba0a7bcf2e2cff88c60d389895a550623.

globskipdots is a new option in bash-5.2 that is also default-on
in that release. The default value is not gated by BASH_COMPAT (see bug #907061),
hence we need to disable it for older Bashes to avoid behaviour changes in ebuilds
and eclasses.

Bug: https://bugs.gentoo.org/907061
Signed-off-by: Sam James <sam@gentoo.org>