Add emerge options to require free space/inodes in tmpdir
This adds
  --jobs-tmpdir-require-free-gb=GB
  --jobs-tmpdir-require-free-kilo-inodes=INODES
as emerge emerge options.

When those are used with --jobs, makes portage/emerge check that
PORTAGE_TMPDIR has sufficient free resources before a new job is
started.

Thanks goes out to Zac Medico, as this was inspired by
https://github.com/gentoo/portage/pull/1351, with the following
differences:

- options are absolute values, not relatives ones
- defaults for both options are specified
- option values are scaled, using a decaying function, considering
  the number or running jobs
- emit a warning once a threshold is reached

Note that the scaling of the resource constraints can not be perfect
in the presence of concurrently running emerge jobs and without
_can_add_job() being provided with the number of jobs that are
potentially added. It is always possible that a emerge job has not yet
used much of the filesystem when we check the remaining filesystem
resources, and at some later point in time uses much more than the
scaling function accounted for it.

Ultimately, there is a trade-off between portage limiting parallelism
needlessly (but still being able to emerge all packages) and portage
failing due to missing resources in PORTAGE_TMPDIR. The chosen
defaults are rather large and most packages use much less filesystem
resources then the scaling function accounts for them. Therefore, the
implemented approach's idea is to lean towards favoring functionality
over parallelism.

Bug: https://bugs.gentoo.org/934382
Signed-off-by: Florian Schmaus <flow@gentoo.org>
Part-of: https://github.com/gentoo/portage/pull/1353
Closes: https://github.com/gentoo/portage/pull/1353
Signed-off-by: Sam James <sam@gentoo.org>