config: Allow a repository to be configured using one of its aliases
Currently, the `[name]` in repos.conf can only match the primary name.
This is inconvenient if a repository wants to change its name without
breaking existing configurations.

This raises the question of whether to then use the primary name or the
alias in the UI and in the vardb. I went with the primary name because
this is presumably the name that the repository actually wants you to
use.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>