install-qa-check.d: try to detect gcc warnings past color
We have some code we use for config-impl-decl which can strip color
codes out via sed. This is always necessary for config.log because if
color exists it will still be there.

The assumption was, likely, that build.log as managed by portage will
also be ansifilter'd by portage, and therefore stripping color is
unnecessary. But in fact, some people do like color and intentionally
avoid stripping it. This rendered the QA check effectively broken.
Instead we should make no assumptions, and explicitly strip this too.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>
Closes: https://github.com/gentoo/portage/pull/1375
Signed-off-by: Sam James <sam@gentoo.org>